+++ info/index.php
@@ -1,7 +1,7 @@
 <?php
 namespace ConnectFour;
 
-define('ROOT', dirname(__DIR__).'/');
+if(!defined('ROOT')) define('ROOT', dirname(__DIR__).'/');
 require_once(ROOT.'config.php');
 
 $info = ['width' => BOARD_WIDTH, 'height' => BOARD_HEIGHT, 'strategies' => STRATEGIES];