+++ play/index.php
@@ -1,12 +1,13 @@
 <?php
 namespace ConnectFour;
 
-define('ROOT', dirname(__DIR__).'/');
+if(!defined('ROOT')) define('ROOT', dirname(__DIR__).'/');
 require_once(ROOT.'lib.php');
 require_once(ROOT.'model/game.php');
 
 $pid = $_GET['pid'];
 $move = $_GET['move'];
+$gameDir = DATA_DIR.$pid;
 
 if (is_null($pid)) {
   Lib\responseError('PID not specified');
@@ -21,12 +22,12 @@ if (!is_numeric($move)) {
   exit;
 }
 $move = (int) $move;
-if(!file_exists(DATA_DIR.$pid)) {
+if(!file_exists($gameDir)) {
   Lib\responseError('Unknown PID');
   exit;
 }
 
-$data = json_decode(file_get_contents(DATA_DIR.$pid), true);
+$data = json_decode(file_get_contents($gameDir), true);
 $strategy = $data['strategy'];
 assert(in_array($strategy, STRATEGIES), "Game data contains unknown strategy.");
 $game = new Model\Game($data['game']);
@@ -38,16 +39,10 @@ if(!in_array($move, $availableMoves)) {
 
 $game->doMove($move, 0);
 
+$ackMove = ['slot'=> $move, 'isWin'=> $game->isWin(), 'isDraw'=> $game->isDraw(), 'row'=> $game->getRow()];
 if($game->isGameOver()) {
-  $winningMove = $game->didLastMoveWin();
-  if($winningMove) {
-    Lib\responseSuccess(['ack_move'=> ['slot'=> $move, 'isWin'=> true, 'isDraw'=> false, 'row'=>$winningMove]]);
-  } elseif($game->isDraw()) {
-    Lib\responseSuccess(['ack_move'=> ['slot'=> $move, 'isWin'=> false, 'isDraw'=> true, 'row'=> []]]);
-  } else {
-    Lib\assert(false, "Game without a draw is over but nobody won.");
-  }
-  unlink(DATA_DIR.$pid);
+  Lib\responseSuccess(['ack_move'=> $ackMove]);
+  unlink($gameDir);
   exit;
 }
 
@@ -57,22 +52,13 @@ assert(!empty($availableMoves), "Game is not over, but there are no available mo
 $computerMove = $availableMoves[rand(0, count($availableMoves)-1)];
 $game->doMove($computerMove, 1);
 
+$responseMove = ['slot'=> $computerMove, 'isWin'=> $game->isWin(), 'isDraw'=> $game->isDraw(), 'row'=> $game->getRow()];
 if($game->isGameOver()) {
-  $winningMove = $game->didLastMoveWin();
-  if($winningMove) {
-    Lib\responseSuccess(['ack_move'=> ['slot'=> $move, 'isWin'=> false, 'isDraw'=> false, 'row'=> []],
-                         'move'=> ['slot'=> $computerMove, 'isWin'=> true, 'isDraw'=> false, 'row'=> $winningMove]]);
-  } elseif($game->isDraw()) {
-    Lib\responseSuccess(['ack_move'=> ['slot'=> $move, 'isWin'=> false, 'isDraw'=> false, 'row'=> []],
-                         'move'=> ['slot'=> $computerMove, 'isWin'=> false, 'isDraw'=> true, 'row'=> []]]);
-  } else {
-    assert(false, "Game without a draw is over but nobody won.");
-  }
-  unlink(DATA_DIR.$pid);
+  Lib\responseSuccess(['ack_move'=> $ackMove, 'move'=> $responseMove]);
+  unlink($gameDir);
   exit;
 }
 
 $contents = json_encode(['strategy'=> $strategy, 'game'=> $game]);
-file_put_contents(DATA_DIR.$pid, $contents);
-Lib\responseSuccess(['ack_move'=> ['slot'=> $move, 'isWin'=> false, 'isDraw'=> false, 'row'=> []],
-                     'move'=> ['slot'=> $computerMove, 'isWin'=> false, 'isDraw'=> false, 'row'=> []]]);
+file_put_contents($gameDir, $contents);
+Lib\responseSuccess(['ack_move'=> $ackMove, 'move'=> $responseMove]);